# chapter03/internal/rest/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Tasks",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "TaskService"
    }
  ],
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/tasks": {
      "get": {
        "summary": "Get All Tasks",
        "operationId": "getAllTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/taskpbGetAllTaskResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Task"
        ]
      },
      "post": {
        "summary": "Create new Task",
        "operationId": "createTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/taskpbCreateTaskResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/taskpbCreateTaskRequest"
            }
          }
        ],
        "tags": [
          "Task"
        ]
      }
    },
    "/api/tasks/{id}/done": {
      "put": {
        "summary": "Mark task as done",
        "operationId": "taskDone",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/taskpbTaskDoneResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "tags": [
          "Task"
        ]
      }
    },
    "/api/tasks/{id}/inprogress": {
      "put": {
        "summary": "Mark task as in progress",
        "operationId": "taskInProgress",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/taskpbTaskInProgressResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "tags": [
          "Task"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "taskpbCreateTaskRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "taskpbCreateTaskResponse": {
      "type": "object",
      "properties": {
        "task": {
          "$ref": "#/definitions/taskpbTask"
        }
      }
    },
    "taskpbGetAllTaskResponse": {
      "type": "object",
      "properties": {
        "tasks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/taskpbTask"
          }
        }
      }
    },
    "taskpbTask": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
    "taskpbTaskDoneResponse": {
      "type": "object",
      "properties": {
        "task": {
          "$ref": "#/definitions/taskpbTask"
        }
      }
    },
    "taskpbTaskInProgressResponse": {
      "type": "object",
      "properties": {
        "task": {
          "$ref": "#/definitions/taskpbTask"
        }
      }
    }
  }
}